python-backend/requirements.txt
exa-py>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
flask>=2.0.0
flask-cors>=3.0.0
anthropic>=0.39.0
pypdf>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
